continue split
