+++ pkg/workflow/compiler.go
@@ -7,9 +7,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/githubnext/gh-aw/pkg/console"
 	"github.com/githubnext/gh-aw/pkg/constants"
@@ -177,26 +175,6 @@ type SafeOutputsConfig struct {
 	GitHubToken                     string                                 `yaml:"github-token,omitempty"`   // GitHub token for safe output jobs
 	MaximumPatchSize                int                                    `yaml:"max-patch-size,omitempty"` // Maximum allowed patch size in KB (defaults to 1024)
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 // CompileWorkflow converts a markdown workflow to GitHub Actions YAML
 func (c *Compiler) CompileWorkflow(markdownPath string) error {
@@ -575,7 +553,7 @@ func (c *Compiler) parseWorkflowFile(markdownPath string) (*WorkflowData, error)
 	}
 
 	// Check if frontmatter specifies a custom name and use it instead
-	frontmatterName := c.extractStringValue(result.Frontmatter, "name")
+	frontmatterName := extractStringValue(result.Frontmatter, "name")
 	if frontmatterName != "" {
 		workflowName = frontmatterName
 	}
@@ -623,7 +601,7 @@ func (c *Compiler) parseWorkflowFile(markdownPath string) (*WorkflowData, error)
 
 	workflowData.Command = c.extractCommandName(result.Frontmatter)
 	workflowData.Jobs = c.extractJobsFromFrontmatter(result.Frontmatter)
-	workflowData.Roles = c.extractRolesPermissions(result.Frontmatter)
+	workflowData.Roles = c.extractRoles(result.Frontmatter)
 
 	// Use the already extracted output configuration
 	workflowData.SafeOutputs = safeOutputs
@@ -635,7 +613,7 @@ func (c *Compiler) parseWorkflowFile(markdownPath string) (*WorkflowData, error)
 	}
 
 	// Apply defaults
-	c.applyDefaults(workflowData, markdownPath, result.Frontmatter)
+	c.applyDefaults(workflowData, markdownPath)
 
 	// Apply pull request draft filter if specified
 	c.applyPullRequestDraftFilter(workflowData, result.Frontmatter)
@@ -644,41 +622,6 @@ func (c *Compiler) parseWorkflowFile(markdownPath string) (*WorkflowData, error)
 	c.applyPullRequestForkFilter(workflowData, result.Frontmatter)
 
 	return workflowData, nil
-}
-
-// extractNetworkPermissions extracts network permissions from frontmatter
-func (c *Compiler) extractNetworkPermissions(frontmatter map[string]any) *NetworkPermissions {
-	if network, exists := frontmatter["network"]; exists {
-		// Handle string format: "defaults"
-		if networkStr, ok := network.(string); ok {
-			if networkStr == "defaults" {
-				return &NetworkPermissions{
-					Mode: "defaults",
-				}
-			}
-			// Unknown string format, return nil
-			return nil
-		}
-
-		// Handle object format: { allowed: [...] } or {}
-		if networkObj, ok := network.(map[string]any); ok {
-			permissions := &NetworkPermissions{}
-
-			// Extract allowed domains if present
-			if allowed, hasAllowed := networkObj["allowed"]; hasAllowed {
-				if allowedSlice, ok := allowed.([]any); ok {
-					for _, domain := range allowedSlice {
-						if domainStr, ok := domain.(string); ok {
-							permissions.Allowed = append(permissions.Allowed, domainStr)
-						}
-					}
-				}
-			}
-			// Empty object {} means no network access (empty allowed list)
-			return permissions
-		}
-	}
-	return nil
 }
 
 // extractTopLevelYAMLSection extracts a top-level YAML section from the frontmatter map
@@ -743,20 +686,6 @@ func (c *Compiler) extractExpressionFromIfString(ifString string) string {
 
 	// Return the string as-is (it's just the expression)
 	return ifString
-}
-
-// extractStringValue extracts a string value from the frontmatter map
-func (c *Compiler) extractStringValue(frontmatter map[string]any, key string) string {
-	value, exists := frontmatter[key]
-	if !exists {
-		return ""
-	}
-
-	if strValue, ok := value.(string); ok {
-		return strValue
-	}
-
-	return ""
 }
 
 // commentOutProcessedFieldsInOnSection comments out draft, fork, and forks fields in pull_request sections within the YAML string
@@ -835,54 +764,6 @@ func (c *Compiler) commentOutProcessedFieldsInOnSection(yamlStr string) string {
 	return strings.Join(result, "\n")
 }
 
-// extractYAMLValue extracts a scalar value from the frontmatter map
-func (c *Compiler) extractYAMLValue(frontmatter map[string]any, key string) string {
-	if value, exists := frontmatter[key]; exists {
-		if str, ok := value.(string); ok {
-			return str
-		}
-		if num, ok := value.(int); ok {
-			return fmt.Sprintf("%d", num)
-		}
-		if num, ok := value.(int64); ok {
-			return fmt.Sprintf("%d", num)
-		}
-		if num, ok := value.(uint64); ok {
-			return fmt.Sprintf("%d", num)
-		}
-		if float, ok := value.(float64); ok {
-			return fmt.Sprintf("%.0f", float)
-		}
-	}
-	return ""
-}
-
-// extractStopAfterFromOn extracts the stop-after value from the on: section
-func (c *Compiler) extractStopAfterFromOn(frontmatter map[string]any) (string, error) {
-	onSection, exists := frontmatter["on"]
-	if !exists {
-		return "", nil
-	}
-
-	// Handle different formats of the on: section
-	switch on := onSection.(type) {
-	case string:
-		// Simple string format like "on: push" - no stop-after possible
-		return "", nil
-	case map[string]any:
-		// Complex object format - look for stop-after
-		if stopAfter, exists := on["stop-after"]; exists {
-			if str, ok := stopAfter.(string); ok {
-				return str, nil
-			}
-			return "", fmt.Errorf("stop-after value must be a string")
-		}
-		return "", nil
-	default:
-		return "", fmt.Errorf("invalid on: section format")
-	}
-}
-
 // parseOnSection parses the "on" section from frontmatter to extract command triggers, reactions, and other events
 func (c *Compiler) parseOnSection(frontmatter map[string]any, workflowData *WorkflowData, markdownPath string) error {
 	// Check if "command" is used as a trigger in the "on" section
@@ -955,50 +836,6 @@ func (c *Compiler) parseOnSection(frontmatter map[string]any, workflowData *Work
 	return nil
 }
 
-// processStopAfterConfiguration extracts and processes stop-after configuration from frontmatter
-func (c *Compiler) processStopAfterConfiguration(frontmatter map[string]any, workflowData *WorkflowData) error {
-	// Extract stop-after from the on: section
-	stopAfter, err := c.extractStopAfterFromOn(frontmatter)
-	if err != nil {
-		return err
-	}
-	workflowData.StopTime = stopAfter
-
-	// Resolve relative stop-after to absolute time if needed
-	if workflowData.StopTime != "" {
-		resolvedStopTime, err := resolveStopTime(workflowData.StopTime, time.Now().UTC())
-		if err != nil {
-			return fmt.Errorf("invalid stop-after format: %w", err)
-		}
-		originalStopTime := stopAfter
-		workflowData.StopTime = resolvedStopTime
-
-		if c.verbose && isRelativeStopTime(originalStopTime) {
-			fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Resolved relative stop-after to: %s", resolvedStopTime)))
-		} else if c.verbose && originalStopTime != resolvedStopTime {
-			fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Parsed absolute stop-after from '%s' to: %s", originalStopTime, resolvedStopTime)))
-		}
-	}
-
-	return nil
-}
-
-// filterMapKeys creates a new map excluding the specified keys
-func filterMapKeys(original map[string]any, excludeKeys ...string) map[string]any {
-	excludeSet := make(map[string]bool)
-	for _, key := range excludeKeys {
-		excludeSet[key] = true
-	}
-
-	result := make(map[string]any)
-	for key, value := range original {
-		if !excludeSet[key] {
-			result[key] = value
-		}
-	}
-	return result
-}
-
 // generateJobName converts a workflow name to a valid YAML job identifier
 func (c *Compiler) generateJobName(workflowName string) string {
 	// Convert to lowercase and replace spaces and special characters with hyphens
@@ -1051,93 +888,8 @@ func (c *Compiler) extractCommandName(frontmatter map[string]any) string {
 	return ""
 }
 
-// extractRolesPermissions extracts the 'roles' field from frontmatter to determine permission requirements
-func (c *Compiler) extractRolesPermissions(frontmatter map[string]any) []string {
-	if rolesValue, exists := frontmatter["roles"]; exists {
-		switch v := rolesValue.(type) {
-		case string:
-			if v == "all" {
-				// Special case: "all" means no restrictions
-				return []string{"all"}
-			}
-			// Single permission level as string
-			return []string{v}
-		case []any:
-			// Array of permission levels
-			var permissions []string
-			for _, item := range v {
-				if str, ok := item.(string); ok {
-					permissions = append(permissions, str)
-				}
-			}
-			return permissions
-		case []string:
-			// Already a string slice
-			return v
-		}
-	}
-	// Default: require admin or maintainer permissions
-	return []string{"admin", "maintainer"}
-}
-
-// hasSafeEventsOnly checks if the workflow uses only safe events that don't require permission checks
-func (c *Compiler) hasSafeEventsOnly(data *WorkflowData, frontmatter map[string]any) bool {
-	// If user explicitly specified "roles: all", skip permission checks
-	if len(data.Roles) == 1 && data.Roles[0] == "all" {
-		return true
-	}
-
-	// Parse the "on" section to determine events
-	if onValue, exists := frontmatter["on"]; exists {
-		if onMap, ok := onValue.(map[string]any); ok {
-			// Check if only safe events are present
-			hasUnsafeEvents := false
-
-			for eventName := range onMap {
-				// Skip command events as they are handled separately
-				// Skip stop-after and reaction as they are not event types
-				if eventName == "command" || eventName == "stop-after" || eventName == "reaction" {
-					continue
-				}
-
-				// Check if this event is in the safe list
-				isSafe := false
-				for _, safeEvent := range constants.SafeWorkflowEvents {
-					if eventName == safeEvent {
-						isSafe = true
-						break
-					}
-				}
-				if !isSafe {
-					hasUnsafeEvents = true
-					break
-				}
-			}
-
-			// If there are events and none are unsafe, then it's safe
-			eventCount := len(onMap)
-			// Subtract non-event entries
-			if _, hasCommand := onMap["command"]; hasCommand {
-				eventCount--
-			}
-			if _, hasStopAfter := onMap["stop-after"]; hasStopAfter {
-				eventCount--
-			}
-			if _, hasReaction := onMap["reaction"]; hasReaction {
-				eventCount--
-			}
-
-			return eventCount > 0 && !hasUnsafeEvents
-		}
-	}
-
-	// If no "on" section or it's a string, check for default command trigger
-	// For command workflows, they are not considered "safe only"
-	return false
-}
-
 // applyDefaults applies default values for missing workflow sections
-func (c *Compiler) applyDefaults(data *WorkflowData, markdownPath string, frontmatter map[string]any) {
+func (c *Compiler) applyDefaults(data *WorkflowData, markdownPath string) {
 	// Check if this is a command trigger workflow (by checking if user specified "on.command")
 	isCommandTrigger := false
 	if data.On == "" {
@@ -1637,30 +1389,6 @@ func (c *Compiler) detectTextOutputUsage(markdownContent string) bool {
 	return hasUsage
 }
 
-// indentYAMLLines adds indentation to all lines of a multi-line YAML string except the first
-func (c *Compiler) indentYAMLLines(yamlContent, indent string) string {
-	if yamlContent == "" {
-		return yamlContent
-	}
-
-	lines := strings.Split(yamlContent, "\n")
-	if len(lines) <= 1 {
-		return yamlContent
-	}
-
-	// First line doesn't get additional indentation
-	result := lines[0]
-	for i := 1; i < len(lines); i++ {
-		if strings.TrimSpace(lines[i]) != "" {
-			result += "\n" + indent + lines[i]
-		} else {
-			result += "\n" + lines[i]
-		}
-	}
-
-	return result
-}
-
 // generateYAML generates the complete GitHub Actions YAML content
 func (c *Compiler) generateYAML(data *WorkflowData, markdownPath string) (string, error) {
 	// Reset job manager for this compilation
@@ -1715,36 +1443,6 @@ func (c *Compiler) generateYAML(data *WorkflowData, markdownPath string) (string
 	return yaml.String(), nil
 }
 
-// needsPermissionChecks determines if the workflow needs permission checks
-func (c *Compiler) needsPermissionChecks(data *WorkflowData) bool {
-	// If user explicitly specified "roles: all", no permission checks needed
-	if len(data.Roles) == 1 && data.Roles[0] == "all" {
-		return false
-	}
-
-	// Permission checks are needed by default unless workflow uses only safe events
-	// Safe events: workflow_dispatch, workflow_run, schedule
-	// For now, we'll implement a simple heuristic since we don't have frontmatter here
-	// We'll implement the full logic later when we have access to frontmatter
-	return true
-}
-
-// needsPermissionChecksWithFrontmatter determines if the workflow needs permission checks with full context
-func (c *Compiler) needsPermissionChecksWithFrontmatter(data *WorkflowData, frontmatter map[string]any) bool {
-	// If user explicitly specified "roles: all", no permission checks needed
-	if len(data.Roles) == 1 && data.Roles[0] == "all" {
-		return false
-	}
-
-	// Check if the workflow uses only safe events (only if frontmatter is available)
-	if frontmatter != nil && c.hasSafeEventsOnly(data, frontmatter) {
-		return false
-	}
-
-	// Permission checks are needed by default for non-safe events
-	return true
-}
-
 // isTaskJobNeeded determines if the task job is required
 func (c *Compiler) isTaskJobNeeded(data *WorkflowData, needsPermissionCheck bool) bool {
 	// Task job is needed if:
@@ -1772,12 +1470,7 @@ func (c *Compiler) buildJobs(data *WorkflowData, markdownPath string) error {
 
 	// Build task job if needed (preamble job that handles runtime conditions and permission checks)
 	var taskJobCreated bool
-	var needsPermissionCheck bool
-	if frontmatter != nil {
-		needsPermissionCheck = c.needsPermissionChecksWithFrontmatter(data, frontmatter)
-	} else {
-		needsPermissionCheck = c.needsPermissionChecks(data)
-	}
+	needsPermissionCheck := c.needsRoleCheck(data, frontmatter)
 
 	if c.isTaskJobNeeded(data, needsPermissionCheck) {
 		taskJob, err := c.buildTaskJob(data, frontmatter)
@@ -1810,122 +1503,134 @@ func (c *Compiler) buildJobs(data *WorkflowData, markdownPath string) error {
 		return fmt.Errorf("failed to add main job: %w", err)
 	}
 
-	if data.SafeOutputs != nil {
-		// Build create_issue job if output.create_issue is configured
-		if data.SafeOutputs.CreateIssues != nil {
-			createIssueJob, err := c.buildCreateOutputIssueJob(data, jobName, taskJobCreated, frontmatter)
-			if err != nil {
-				return fmt.Errorf("failed to build create_issue job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createIssueJob); err != nil {
-				return fmt.Errorf("failed to add create_issue job: %w", err)
-			}
-		}
-
-		// Build create_discussion job if output.create_discussion is configured
-		if data.SafeOutputs.CreateDiscussions != nil {
-			createDiscussionJob, err := c.buildCreateOutputDiscussionJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build create_discussion job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createDiscussionJob); err != nil {
-				return fmt.Errorf("failed to add create_discussion job: %w", err)
-			}
-		}
-
-		// Build create_issue_comment job if output.add-comment is configured
-		if data.SafeOutputs.AddComments != nil {
-			createCommentJob, err := c.buildCreateOutputAddCommentJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build create_issue_comment job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createCommentJob); err != nil {
-				return fmt.Errorf("failed to add create_issue_comment job: %w", err)
-			}
-		}
-
-		// Build create_pr_review_comment job if output.create-pull-request-review-comment is configured
-		if data.SafeOutputs.CreatePullRequestReviewComments != nil {
-			createPRReviewCommentJob, err := c.buildCreateOutputPullRequestReviewCommentJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build create_pr_review_comment job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createPRReviewCommentJob); err != nil {
-				return fmt.Errorf("failed to add create_pr_review_comment job: %w", err)
-			}
-		}
-
-		// Build create_code_scanning_alert job if output.create-code-scanning-alert is configured
-		if data.SafeOutputs.CreateCodeScanningAlerts != nil {
-			// Extract the workflow filename without extension for rule ID prefix
-			workflowFilename := strings.TrimSuffix(filepath.Base(markdownPath), ".md")
-			createCodeScanningAlertJob, err := c.buildCreateOutputCodeScanningAlertJob(data, jobName, workflowFilename)
-			if err != nil {
-				return fmt.Errorf("failed to build create_code_scanning_alert job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createCodeScanningAlertJob); err != nil {
-				return fmt.Errorf("failed to add create_code_scanning_alert job: %w", err)
-			}
-		}
-
-		// Build create_pull_request job if output.create-pull-request is configured
-		if data.SafeOutputs.CreatePullRequests != nil {
-			createPullRequestJob, err := c.buildCreateOutputPullRequestJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build create_pull_request job: %w", err)
-			}
-			if err := c.jobManager.AddJob(createPullRequestJob); err != nil {
-				return fmt.Errorf("failed to add create_pull_request job: %w", err)
-			}
-		}
-
-		// Build add_labels job if output.add-labels is configured (including null/empty)
-		if data.SafeOutputs.AddLabels != nil {
-			addLabelsJob, err := c.buildCreateOutputLabelJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build add_labels job: %w", err)
-			}
-			if err := c.jobManager.AddJob(addLabelsJob); err != nil {
-				return fmt.Errorf("failed to add add_labels job: %w", err)
-			}
-		}
-
-		// Build update_issue job if output.update-issue is configured
-		if data.SafeOutputs.UpdateIssues != nil {
-			updateIssueJob, err := c.buildCreateOutputUpdateIssueJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build update_issue job: %w", err)
-			}
-			if err := c.jobManager.AddJob(updateIssueJob); err != nil {
-				return fmt.Errorf("failed to add update_issue job: %w", err)
-			}
-		}
-
-		// Build push_to_pr_branch job if output.push-to-pr-branch is configured
-		if data.SafeOutputs.PushToPullRequestBranch != nil {
-			pushToBranchJob, err := c.buildCreateOutputPushToPullRequestBranchJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build push_to_pr_branch job: %w", err)
-			}
-			if err := c.jobManager.AddJob(pushToBranchJob); err != nil {
-				return fmt.Errorf("failed to add push_to_pr_branch job: %w", err)
-			}
-		}
-
-		// Build missing_tool job (always enabled when SafeOutputs exists)
-		if data.SafeOutputs.MissingTool != nil {
-			missingToolJob, err := c.buildCreateOutputMissingToolJob(data, jobName)
-			if err != nil {
-				return fmt.Errorf("failed to build missing_tool job: %w", err)
-			}
-			if err := c.jobManager.AddJob(missingToolJob); err != nil {
-				return fmt.Errorf("failed to add missing_tool job: %w", err)
-			}
-		}
+	// Build safe outputs jobs if configured
+	if err := c.buildSafeOutputsJobs(data, jobName, taskJobCreated, frontmatter, markdownPath); err != nil {
+		return fmt.Errorf("failed to build safe outputs jobs: %w", err)
 	}
 	// Build additional custom jobs from frontmatter jobs section
 	if err := c.buildCustomJobs(data); err != nil {
 		return fmt.Errorf("failed to build custom jobs: %w", err)
+	}
+
+	return nil
+}
+
+// buildSafeOutputsJobs creates all safe outputs jobs if configured
+func (c *Compiler) buildSafeOutputsJobs(data *WorkflowData, jobName string, taskJobCreated bool, frontmatter map[string]any, markdownPath string) error {
+	if data.SafeOutputs == nil {
+		return nil
+	}
+
+	// Build create_issue job if output.create_issue is configured
+	if data.SafeOutputs.CreateIssues != nil {
+		createIssueJob, err := c.buildCreateOutputIssueJob(data, jobName, taskJobCreated, frontmatter)
+		if err != nil {
+			return fmt.Errorf("failed to build create_issue job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createIssueJob); err != nil {
+			return fmt.Errorf("failed to add create_issue job: %w", err)
+		}
+	}
+
+	// Build create_discussion job if output.create_discussion is configured
+	if data.SafeOutputs.CreateDiscussions != nil {
+		createDiscussionJob, err := c.buildCreateOutputDiscussionJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build create_discussion job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createDiscussionJob); err != nil {
+			return fmt.Errorf("failed to add create_discussion job: %w", err)
+		}
+	}
+
+	// Build create_issue_comment job if output.add-comment is configured
+	if data.SafeOutputs.AddComments != nil {
+		createCommentJob, err := c.buildCreateOutputAddCommentJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build create_issue_comment job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createCommentJob); err != nil {
+			return fmt.Errorf("failed to add create_issue_comment job: %w", err)
+		}
+	}
+
+	// Build create_pr_review_comment job if output.create-pull-request-review-comment is configured
+	if data.SafeOutputs.CreatePullRequestReviewComments != nil {
+		createPRReviewCommentJob, err := c.buildCreateOutputPullRequestReviewCommentJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build create_pr_review_comment job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createPRReviewCommentJob); err != nil {
+			return fmt.Errorf("failed to add create_pr_review_comment job: %w", err)
+		}
+	}
+
+	// Build create_code_scanning_alert job if output.create-code-scanning-alert is configured
+	if data.SafeOutputs.CreateCodeScanningAlerts != nil {
+		// Extract the workflow filename without extension for rule ID prefix
+		workflowFilename := strings.TrimSuffix(filepath.Base(markdownPath), ".md")
+		createCodeScanningAlertJob, err := c.buildCreateOutputCodeScanningAlertJob(data, jobName, workflowFilename)
+		if err != nil {
+			return fmt.Errorf("failed to build create_code_scanning_alert job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createCodeScanningAlertJob); err != nil {
+			return fmt.Errorf("failed to add create_code_scanning_alert job: %w", err)
+		}
+	}
+
+	// Build create_pull_request job if output.create-pull-request is configured
+	if data.SafeOutputs.CreatePullRequests != nil {
+		createPullRequestJob, err := c.buildCreateOutputPullRequestJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build create_pull_request job: %w", err)
+		}
+		if err := c.jobManager.AddJob(createPullRequestJob); err != nil {
+			return fmt.Errorf("failed to add create_pull_request job: %w", err)
+		}
+	}
+
+	// Build add_labels job if output.add-labels is configured (including null/empty)
+	if data.SafeOutputs.AddLabels != nil {
+		addLabelsJob, err := c.buildCreateOutputLabelJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build add_labels job: %w", err)
+		}
+		if err := c.jobManager.AddJob(addLabelsJob); err != nil {
+			return fmt.Errorf("failed to add add_labels job: %w", err)
+		}
+	}
+
+	// Build update_issue job if output.update-issue is configured
+	if data.SafeOutputs.UpdateIssues != nil {
+		updateIssueJob, err := c.buildCreateOutputUpdateIssueJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build update_issue job: %w", err)
+		}
+		if err := c.jobManager.AddJob(updateIssueJob); err != nil {
+			return fmt.Errorf("failed to add update_issue job: %w", err)
+		}
+	}
+
+	// Build push_to_pr_branch job if output.push-to-pr-branch is configured
+	if data.SafeOutputs.PushToPullRequestBranch != nil {
+		pushToBranchJob, err := c.buildCreateOutputPushToPullRequestBranchJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build push_to_pr_branch job: %w", err)
+		}
+		if err := c.jobManager.AddJob(pushToBranchJob); err != nil {
+			return fmt.Errorf("failed to add push_to_pr_branch job: %w", err)
+		}
+	}
+
+	// Build missing_tool job (always enabled when SafeOutputs exists)
+	if data.SafeOutputs.MissingTool != nil {
+		missingToolJob, err := c.buildCreateOutputMissingToolJob(data, jobName)
+		if err != nil {
+			return fmt.Errorf("failed to build missing_tool job: %w", err)
+		}
+		if err := c.jobManager.AddJob(missingToolJob); err != nil {
+			return fmt.Errorf("failed to add missing_tool job: %w", err)
+		}
 	}
 
 	return nil
@@ -1937,14 +1642,9 @@ func (c *Compiler) buildTaskJob(data *WorkflowData, frontmatter map[string]any) 
 	var steps []string
 
 	// Add team member check based on new permission requirements (issue #567)
-	var needsPermissionCheck bool
-	if frontmatter != nil {
-		needsPermissionCheck = c.needsPermissionChecksWithFrontmatter(data, frontmatter)
-	} else {
-		needsPermissionCheck = c.needsPermissionChecks(data)
-	}
+	needsRoleCheck := c.needsRoleCheck(data, frontmatter)
 
-	if needsPermissionCheck || data.Command != "" {
+	if needsRoleCheck || data.Command != "" {
 		if data.Command != "" {
 			steps = append(steps, "      - name: Check team membership for command workflow\n")
 		} else {
@@ -1961,7 +1661,7 @@ func (c *Compiler) buildTaskJob(data *WorkflowData, frontmatter map[string]any) 
 		steps = append(steps, "          script: |\n")
 
 		// Generate the JavaScript code for the permission check
-		scriptContent := c.generatePermissionCheckScript(data.Roles)
+		scriptContent := c.generateRoleCheckScript(data.Roles)
 		scriptLines := strings.Split(scriptContent, "\n")
 		for _, line := range scriptLines {
 			if strings.TrimSpace(line) != "" {
@@ -2003,7 +1703,7 @@ func (c *Compiler) buildTaskJob(data *WorkflowData, frontmatter map[string]any) 
 
 	// Add actions: write permission if team member checks are present
 	// Any workflow that needs permission checks will use setCancelled() which requires actions: write
-	requiresWorkflowCancellation := data.Command != "" || needsPermissionCheck
+	requiresWorkflowCancellation := data.Command != "" || needsRoleCheck
 
 	if requiresWorkflowCancellation {
 		job.Permissions = "permissions:\n      actions: write  # Required for github.rest.actions.cancelWorkflowRun()"
@@ -2012,126 +1712,15 @@ func (c *Compiler) buildTaskJob(data *WorkflowData, frontmatter map[string]any) 
 	return job, nil
 }
 
-// generatePermissionCheckScript generates JavaScript code to check user permissions
-func (c *Compiler) generatePermissionCheckScript(requiredPermissions []string) string {
-	// If "all" is specified, no checks needed (this shouldn't happen since needsPermissionChecks would return false)
-	if len(requiredPermissions) == 1 && requiredPermissions[0] == "all" {
-		return `
-core.setOutput("is_team_member", "true");
-console.log("Permission check skipped - 'roles: all' specified");`
-	}
-
-	// Use the embedded check_permissions.cjs script
-	// The GITHUB_AW_REQUIRED_ROLES environment variable is set via the env field
-	return checkPermissionsScript
-}
-
-// buildAddReactionJob creates the add_reaction job
-func (c *Compiler) buildAddReactionJob(data *WorkflowData, taskJobCreated bool, frontmatter map[string]any) (*Job, error) {
-	reactionCondition := buildReactionCondition()
-
-	var steps []string
-
-	// Add permission checks if no task job was created but permission checks are needed
-	if !taskJobCreated && c.needsPermissionChecks(data) {
-		// Add team member check step
-		steps = append(steps, "      - name: Check team membership for workflow\n")
-		steps = append(steps, "        id: check-team-member\n")
-		steps = append(steps, "        uses: actions/github-script@v8\n")
-
-		// Add environment variables for permission check
-		steps = append(steps, "        env:\n")
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_REQUIRED_ROLES: %s\n", strings.Join(data.Roles, ",")))
-
-		steps = append(steps, "        with:\n")
-		steps = append(steps, "          script: |\n")
-
-		// Generate the JavaScript code for the permission check
-		scriptContent := c.generatePermissionCheckScript(data.Roles)
-		scriptLines := strings.Split(scriptContent, "\n")
-		for _, line := range scriptLines {
-			if strings.TrimSpace(line) != "" {
-				steps = append(steps, fmt.Sprintf("            %s\n", line))
-			}
-		}
-	}
-
-	steps = append(steps, fmt.Sprintf("      - name: Add %s reaction to the triggering item\n", data.AIReaction))
-	steps = append(steps, "        id: react\n")
-	steps = append(steps, "        uses: actions/github-script@v8\n")
-
-	// Add environment variables
-	steps = append(steps, "        env:\n")
-	steps = append(steps, fmt.Sprintf("          GITHUB_AW_REACTION: %s\n", data.AIReaction))
-	if data.Command != "" {
-		steps = append(steps, fmt.Sprintf("          GITHUB_AW_COMMAND: %s\n", data.Command))
-	}
-
-	steps = append(steps, "        with:\n")
-	steps = append(steps, "          script: |\n")
-
-	// Add each line of the script with proper indentation
-	formattedScript := FormatJavaScriptForYAML(addReactionAndEditCommentScript)
-	steps = append(steps, formattedScript...)
-
-	outputs := map[string]string{
-		"reaction_id": "${{ steps.react.outputs.reaction-id }}",
-	}
-
-	var depends []string
-	if taskJobCreated {
-		depends = []string{"task"} // Depend on the task job only if it exists
-	}
-
-	// Set base permissions
-	permissions := "permissions:\n      issues: write\n      pull-requests: write"
-
-	// Add actions: write permission if team member checks are present for command workflows
-	_, hasExplicitRoles := frontmatter["roles"]
-	requiresWorkflowCancellation := data.Command != "" ||
-		(!taskJobCreated && c.needsPermissionChecks(data) && hasExplicitRoles)
-
-	if requiresWorkflowCancellation {
-		permissions = "permissions:\n      actions: write  # Required for github.rest.actions.cancelWorkflowRun()\n      issues: write\n      pull-requests: write\n      contents: read"
-	}
-
-	job := &Job{
-		Name:        "add_reaction",
-		If:          reactionCondition.Render(),
-		RunsOn:      "runs-on: ubuntu-latest",
-		Permissions: permissions,
-		Steps:       steps,
-		Outputs:     outputs,
-		Needs:       depends,
-	}
-
-	return job, nil
-}
-
-
-
-
-
-
-
-
-
-
 // buildMainJob creates the main workflow job
 func (c *Compiler) buildMainJob(data *WorkflowData, jobName string, taskJobCreated bool, frontmatter map[string]any) (*Job, error) {
 	var steps []string
 
 	// Add permission checks if no task job was created but permission checks are needed
 	if !taskJobCreated {
-		var needsPermissionCheck bool
-		// Check if permission checks are needed using frontmatter if available
-		if frontmatter != nil {
-			needsPermissionCheck = c.needsPermissionChecksWithFrontmatter(data, frontmatter)
-		} else {
-			needsPermissionCheck = c.needsPermissionChecks(data)
-		}
+		needsRoleCheck := c.needsRoleCheck(data, frontmatter)
 
-		if needsPermissionCheck {
+		if needsRoleCheck {
 			// Add team member check step
 			steps = append(steps, "      - name: Check team membership for workflow\n")
 			steps = append(steps, "        id: check-team-member\n")
@@ -2144,8 +1733,8 @@ func (c *Compiler) buildMainJob(data *WorkflowData, jobName string, taskJobCreat
 			steps = append(steps, "        with:\n")
 			steps = append(steps, "          script: |\n")
 
-			// Generate the JavaScript code for the permission check
-			scriptContent := c.generatePermissionCheckScript(data.Roles)
+			// Generate the JavaScript code for the role check
+			scriptContent := c.generateRoleCheckScript(data.Roles)
 			scriptLines := strings.Split(scriptContent, "\n")
 			for _, line := range scriptLines {
 				if strings.TrimSpace(line) != "" {
@@ -2248,272 +1837,6 @@ func (c *Compiler) generateSafetyChecks(yaml *strings.Builder, data *WorkflowDat
 	yaml.WriteString("          echo \"All safety checks passed. Proceeding with agentic tool execution.\"\n")
 	yaml.WriteString("        env:\n")
 	yaml.WriteString("          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}\n")
-}
-
-// generateGitConfiguration generates standardized git credential setup
-func (c *Compiler) generateGitConfiguration(yaml *strings.Builder, data *WorkflowData) {
-	steps := c.generateGitConfigurationSteps()
-	for _, step := range steps {
-		yaml.WriteString(step)
-	}
-}
-
-// generateGitConfigurationSteps generates standardized git credential setup as string steps
-func (c *Compiler) generateGitConfigurationSteps() []string {
-	return []string{
-		"      - name: Configure Git credentials\n",
-		"        run: |\n",
-		"          git config --global user.email \"github-actions[bot]@users.noreply.github.com\"\n",
-		"          git config --global user.name \"${{ github.workflow }}\"\n",
-		"          echo \"Git configured with standard GitHub Actions identity\"\n",
-	}
-}
-
-// generateMCPSetup generates the MCP server configuration setup
-func (c *Compiler) generateMCPSetup(yaml *strings.Builder, tools map[string]any, engine CodingAgentEngine, workflowData *WorkflowData) {
-	// Collect tools that need MCP server configuration
-	var mcpTools []string
-	var proxyTools []string
-
-	// Check if workflowData is valid before accessing its fields
-	if workflowData == nil {
-		return
-	}
-
-	workflowTools := workflowData.Tools
-
-	for toolName, toolValue := range workflowTools {
-		// Standard MCP tools
-		if toolName == "github" || toolName == "playwright" || toolName == "cache-memory" {
-			mcpTools = append(mcpTools, toolName)
-		} else if mcpConfig, ok := toolValue.(map[string]any); ok {
-			// Check if it's explicitly marked as MCP type in the new format
-			if hasMcp, _ := hasMCPConfig(mcpConfig); hasMcp {
-				mcpTools = append(mcpTools, toolName)
-
-				// Check if this tool needs proxy
-				if needsProxySetup, _ := needsProxy(mcpConfig); needsProxySetup {
-					proxyTools = append(proxyTools, toolName)
-				}
-			}
-		}
-	}
-
-	// Check if safe-outputs is enabled and add to MCP tools
-	if workflowData.SafeOutputs != nil && HasSafeOutputsEnabled(workflowData.SafeOutputs) {
-		mcpTools = append(mcpTools, "safe-outputs")
-	}
-
-	// Sort tools to ensure stable code generation
-	sort.Strings(mcpTools)
-	sort.Strings(proxyTools)
-
-	// Generate proxy configuration files inline for proxy-enabled tools
-	// These files will be used automatically by docker compose when MCP tools run
-	if len(proxyTools) > 0 {
-		yaml.WriteString("      - name: Setup Proxy Configuration for MCP Network Restrictions\n")
-		yaml.WriteString("        run: |\n")
-		yaml.WriteString("          echo \"Generating proxy configuration files for MCP tools with network restrictions...\"\n")
-		yaml.WriteString("          \n")
-
-		// Generate proxy configurations inline for each proxy-enabled tool
-		for _, toolName := range proxyTools {
-			if toolConfig, ok := tools[toolName].(map[string]any); ok {
-				c.generateInlineProxyConfig(yaml, toolName, toolConfig)
-			}
-		}
-
-		yaml.WriteString("          echo \"Proxy configuration files generated.\"\n")
-
-		// Pre-pull images and start squid proxy ahead of time to avoid timeouts
-		yaml.WriteString("      - name: Pre-pull images and start Squid proxy\n")
-		yaml.WriteString("        run: |\n")
-		yaml.WriteString("          set -e\n")
-		yaml.WriteString("          echo 'Pre-pulling Docker images for proxy-enabled MCP tools...'\n")
-		yaml.WriteString("          docker pull ubuntu/squid:latest\n")
-
-		// Pull each tool's container image if specified, and bring up squid service
-		for _, toolName := range proxyTools {
-			if toolConfig, ok := tools[toolName].(map[string]any); ok {
-				if mcpConf, err := getMCPConfig(toolConfig, toolName); err == nil {
-					if containerVal, hasContainer := mcpConf["container"]; hasContainer {
-						if containerStr, ok := containerVal.(string); ok && containerStr != "" {
-							fmt.Fprintf(yaml, "          echo 'Pulling %s for tool %s'\n", containerStr, toolName)
-							fmt.Fprintf(yaml, "          docker pull %s\n", containerStr)
-						}
-					}
-				}
-				fmt.Fprintf(yaml, "          echo 'Starting squid-proxy service for %s'\n", toolName)
-				fmt.Fprintf(yaml, "          docker compose -f docker-compose-%s.yml up -d squid-proxy\n", toolName)
-
-				// Enforce that egress from this tool's network can only reach the Squid proxy
-				subnetCIDR, squidIP, _ := computeProxyNetworkParams(toolName)
-				fmt.Fprintf(yaml, "          echo 'Enforcing egress to proxy for %s (subnet %s, squid %s)'\n", toolName, subnetCIDR, squidIP)
-				yaml.WriteString("          if command -v sudo >/dev/null 2>&1; then SUDO=sudo; else SUDO=; fi\n")
-				// Accept established/related connections first (position 1)
-				yaml.WriteString("          $SUDO iptables -C DOCKER-USER -m conntrack --ctstate ESTABLISHED,RELATED -j ACCEPT 2>/dev/null || $SUDO iptables -I DOCKER-USER 1 -m conntrack --ctstate ESTABLISHED,RELATED -j ACCEPT\n")
-				// Accept all egress from Squid IP (position 2)
-				fmt.Fprintf(yaml, "          $SUDO iptables -C DOCKER-USER -s %s -j ACCEPT 2>/dev/null || $SUDO iptables -I DOCKER-USER 2 -s %s -j ACCEPT\n", squidIP, squidIP)
-				// Allow traffic to squid:3128 from the subnet (position 3)
-				fmt.Fprintf(yaml, "          $SUDO iptables -C DOCKER-USER -s %s -d %s -p tcp --dport 3128 -j ACCEPT 2>/dev/null || $SUDO iptables -I DOCKER-USER 3 -s %s -d %s -p tcp --dport 3128 -j ACCEPT\n", subnetCIDR, squidIP, subnetCIDR, squidIP)
-				// Then reject all other egress from that subnet (append to end)
-				fmt.Fprintf(yaml, "          $SUDO iptables -C DOCKER-USER -s %s -j REJECT 2>/dev/null || $SUDO iptables -A DOCKER-USER -s %s -j REJECT\n", subnetCIDR, subnetCIDR)
-			}
-		}
-	}
-
-	// If no MCP tools, no configuration needed
-	if len(mcpTools) == 0 {
-		return
-	}
-
-	// Write safe-outputs MCP server if enabled
-	hasSafeOutputs := workflowData != nil && workflowData.SafeOutputs != nil && HasSafeOutputsEnabled(workflowData.SafeOutputs)
-	if hasSafeOutputs {
-		yaml.WriteString("      - name: Setup Safe Outputs Collector MCP\n")
-		safeOutputConfig := c.generateSafeOutputsConfig(workflowData)
-		if safeOutputConfig != "" {
-			// Add environment variables for JSONL validation
-			yaml.WriteString("        env:\n")
-			fmt.Fprintf(yaml, "          GITHUB_AW_SAFE_OUTPUTS_CONFIG: %q\n", safeOutputConfig)
-		}
-		yaml.WriteString("        run: |\n")
-		yaml.WriteString("          mkdir -p /tmp/safe-outputs\n")
-		yaml.WriteString("          cat > /tmp/safe-outputs/mcp-server.cjs << 'EOF'\n")
-		// Embed the safe-outputs MCP server script
-		for _, line := range FormatJavaScriptForYAML(safeOutputsMCPServerScript) {
-			yaml.WriteString(line)
-		}
-		yaml.WriteString("          EOF\n")
-		yaml.WriteString("          chmod +x /tmp/safe-outputs/mcp-server.cjs\n")
-		yaml.WriteString("          \n")
-	}
-
-	// Use the engine's RenderMCPConfig method
-	yaml.WriteString("      - name: Setup MCPs\n")
-	if hasSafeOutputs {
-		safeOutputConfig := c.generateSafeOutputsConfig(workflowData)
-		if safeOutputConfig != "" {
-			// Add environment variables for JSONL validation
-			yaml.WriteString("        env:\n")
-			fmt.Fprintf(yaml, "          GITHUB_AW_SAFE_OUTPUTS: ${{ env.GITHUB_AW_SAFE_OUTPUTS }}\n")
-			fmt.Fprintf(yaml, "          GITHUB_AW_SAFE_OUTPUTS_CONFIG: %q\n", safeOutputConfig)
-		}
-	}
-	yaml.WriteString("        run: |\n")
-	yaml.WriteString("          mkdir -p /tmp/mcp-config\n")
-	engine.RenderMCPConfig(yaml, tools, mcpTools, workflowData)
-}
-
-func getGitHubDockerImageVersion(githubTool any) string {
-	githubDockerImageVersion := "sha-09deac4" // Default Docker image version
-	// Extract docker_image_version setting from tool properties
-	if toolConfig, ok := githubTool.(map[string]any); ok {
-		if versionSetting, exists := toolConfig["docker_image_version"]; exists {
-			if stringValue, ok := versionSetting.(string); ok {
-				githubDockerImageVersion = stringValue
-			}
-		}
-	}
-	return githubDockerImageVersion
-}
-
-func getPlaywrightDockerImageVersion(playwrightTool any) string {
-	playwrightDockerImageVersion := "latest" // Default Playwright Docker image version
-	// Extract docker_image_version setting from tool properties
-	if toolConfig, ok := playwrightTool.(map[string]any); ok {
-		if versionSetting, exists := toolConfig["docker_image_version"]; exists {
-			if stringValue, ok := versionSetting.(string); ok {
-				playwrightDockerImageVersion = stringValue
-			}
-		}
-	}
-	return playwrightDockerImageVersion
-}
-
-// ensureLocalhostDomainsWorkflow ensures that localhost and 127.0.0.1 are always included
-// in the allowed domains list for Playwright, even when custom domains are specified
-func ensureLocalhostDomainsWorkflow(domains []string) []string {
-	hasLocalhost := false
-	hasLoopback := false
-
-	for _, domain := range domains {
-		if domain == "localhost" {
-			hasLocalhost = true
-		}
-		if domain == "127.0.0.1" {
-			hasLoopback = true
-		}
-	}
-
-	result := make([]string, 0, len(domains)+2)
-
-	// Always add localhost domains first
-	if !hasLocalhost {
-		result = append(result, "localhost")
-	}
-	if !hasLoopback {
-		result = append(result, "127.0.0.1")
-	}
-
-	// Add the rest of the domains
-	result = append(result, domains...)
-
-	return result
-}
-
-// generatePlaywrightAllowedDomains extracts domain list from Playwright tool configuration with bundle resolution
-// Uses the same domain bundle resolution as top-level network configuration, defaulting to localhost only
-func generatePlaywrightAllowedDomains(playwrightTool any, networkPermissions *NetworkPermissions) []string {
-	// Default to localhost only (same as Copilot agent default)
-	allowedDomains := []string{"localhost", "127.0.0.1"}
-
-	// Extract allowed_domains from Playwright tool configuration
-	if toolConfig, ok := playwrightTool.(map[string]any); ok {
-		if domainsConfig, exists := toolConfig["allowed_domains"]; exists {
-			// Create a mock NetworkPermissions structure to use the same domain resolution logic
-			playwrightNetwork := &NetworkPermissions{}
-
-			switch domains := domainsConfig.(type) {
-			case []string:
-				playwrightNetwork.Allowed = domains
-			case []any:
-				// Convert []any to []string
-				allowedDomainsSlice := make([]string, len(domains))
-				for i, domain := range domains {
-					if domainStr, ok := domain.(string); ok {
-						allowedDomainsSlice[i] = domainStr
-					}
-				}
-				playwrightNetwork.Allowed = allowedDomainsSlice
-			case string:
-				// Single domain as string
-				playwrightNetwork.Allowed = []string{domains}
-			}
-
-			// Use the same domain bundle resolution as the top-level network configuration
-			resolvedDomains := GetAllowedDomains(playwrightNetwork)
-
-			// Ensure localhost domains are always included
-			allowedDomains = ensureLocalhostDomainsWorkflow(resolvedDomains)
-		}
-	}
-
-	return allowedDomains
-}
-
-// PlaywrightDockerArgs represents the common Docker arguments for Playwright container
-type PlaywrightDockerArgs struct {
-	ImageVersion   string
-	AllowedDomains []string
-}
-
-// generatePlaywrightDockerArgs creates the common Docker arguments for Playwright MCP server
-func generatePlaywrightDockerArgs(playwrightTool any, networkPermissions *NetworkPermissions) PlaywrightDockerArgs {
-	return PlaywrightDockerArgs{
-		ImageVersion:   getPlaywrightDockerImageVersion(playwrightTool),
-		AllowedDomains: generatePlaywrightAllowedDomains(playwrightTool, networkPermissions),
-	}
 }
 
 // generateMainJobSteps generates the steps section for the main job
@@ -3358,8 +2681,6 @@ func (c *Compiler) extractSafeOutputsConfig(frontmatter map[string]any) *SafeOut
 	return config
 }
 
-
-
 // extractCacheMemoryConfig extracts cache-memory configuration from tools section
 func (c *Compiler) extractCacheMemoryConfig(tools map[string]any) *CacheMemoryConfig {
 	cacheMemoryValue, exists := tools["cache-memory"]
@@ -3416,26 +2737,6 @@ func (c *Compiler) extractCacheMemoryConfig(tools map[string]any) *CacheMemoryCo
 
 	return nil
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 // buildCustomJobs creates custom jobs defined in the frontmatter jobs section
 func (c *Compiler) buildCustomJobs(data *WorkflowData) error {
